共通のレイアウト部分を追加
