+++ frontend/src/App.js
@@ -1,38 +1,33 @@
-import { createBrowserRouter, RouterProvider, Link } from 'react-router-dom';
+import React from 'react';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import Home from './pages/Home';
 import Chat from './pages/Chat';
 import StampRally from './pages/StampRally';
+import MainLayout from './layout/MainLayout';
 
 // 1. ルーティングの定義
+// Layoutコンポーネントを親ルート ("/") に設定し、
+// 各ページをその子ルート (children) として定義します。
+// これにより、Layout内の <Outlet /> の部分に子ページが表示されます。
 const router = createBrowserRouter([
   {
     path: "/",
-    element: (
-      <div>
-        {/* ここにページの中身が表示される */}
-        <Home />
-      </div>
-    ),
+    element: <MainLayout />, // ここで共通レイアウトを指定
+    children: [
+      {
+        index: true, // パスが "/" の時はHomeを表示
+        element: <Home />,
+      },
+      {
+        path: "chat", // "/chat"
+        element: <Chat />,
+      },
+      {
+        path: "stampRally", // "/stampRally"
+        element: <StampRally />,
+      },
+    ],
   },
-  {
-    path: "/chat",
-    element: (
-      <div>
-        {/* ここにページの中身が表示される */}
-        <Chat />
-      </div>
-    ),
-  },
-  {
-    path: "/stampRally",
-    element: (
-      <div>
-        {/* ここにページの中身が表示される */}
-        <StampRally />
-      </div>
-    ),
-  },
-  // ページがあれば追加していく
 ]);
 
 // 2. アプリ全体にルーターを適用
